#1590 Move TrackballControls to libs
