+++ app/assets/javascripts/libs/trackball_controls.js
